requirements.txt
streamlit
tensorflow==2.20.0
scikit-image
opencv-python
numpy
Pillow
requests